refactor(plugin-text-classification): rename text-classificatier to intent-classifier
